Add source line number evaluation
Often with "line(tr)", it's possible to find out on which line in the raw source file a text content segment begins. The text content must be unique for this to work. Together with a URI to the raw source file, the exact origin of the information can be cross-referenced.